"exonsByOverlaps" and "transcriptsByOverlaps" methods for EnsDb objects now use the same 'maxgap' and 'minoverlap' defaults as their generic function.
Default values for 'maxgap' and 'minoverlap' arguments have changed in the
IRanges package for findOverlaps() and family (although this change does
NOT affect the default behavior of findOverlaps() and family, except on
zero-width ranges). As a consequence of this change, the default values
for 'maxgap' and 'minoverlap' needed to be changed in the GenomicFeatures
package for the *ByOverlaps() generics. All corresponding methods also need
to be modified to use these new defaults.