Turn regular function proteinToGenome() into S4 method
In BioC 3.15 proteinToGenome() became an S4 generic function defined in
the GenomicFeatures package, with two methods: a default method (that
will typically operate on a TxDb object), and a method that operates on
a GRangesList object.

Following this change, the proteinToGenome() function defined in the
ensembldb package needs to be turned into an S4 method that operates
on an EnsDb object.